% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_biscuiteer.R
\name{benchmark_biscuiteer}
\alias{benchmark_biscuiteer}
\title{Benchmark iscream::query_all on varying region, thread and file counts}
\usage{
benchmark_biscuiteer(
  bedfiles,
  regions,
  merged,
  min_iterations = 3,
  max_iterations = 5,
  n_regions = 10000,
  n_files = c(1, 10, 50, 100),
  n_threads = c(1, 16),
  outfile = NULL
)
}
\arguments{
\item{bedfiles}{A vector of bedfiles to run tests on}

\item{regions}{A vector of regions to query from}

\item{merged}{Whether the files are CG merged}

\item{min_iterations}{The min iterations for bench::mark to run}

\item{max_iterations}{The max iterations for bench::mark to run}

\item{n_regions}{A vector of region counts to benchmark}

\item{n_files}{A vector of file counts to benchmark}

\item{n_threads}{A vector of thread counts to benchmark}

\item{outfile}{Optional file to write the benchmark to}

\item{threads}{The number of threads to run on}
}
\description{
Benchmark iscream::query_all on varying region, thread and file counts
}
