% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_tabix.R
\name{benchmark_tabix}
\alias{benchmark_tabix}
\title{Benchmark iscream::tabix against Rsamtools::scanTabix}
\usage{
benchmark_tabix(
  bedfile,
  regions,
  min_iterations = 5,
  max_iterations = 10,
  raw = TRUE,
  n_regions = c(1, 100, 500, 1000, 5000, 10000, 30000),
  outfile = NULL
)
}
\arguments{
\item{bedfile}{A bedfile to run tests on}

\item{regions}{A vector of regions to query from}

\item{min_iterations}{The min iterations for bench::mark to run}

\item{max_iterations}{The max iterations for bench::mark to run}

\item{raw}{Whether to return raw strings in a named list as Rsamtools::scanTabix does}

\item{n_regions}{A vector of region counts to benchmark}

\item{outfile}{Optional file to write the benchmark to}
}
\description{
Benchmark iscream::tabix against Rsamtools::scanTabix
}
