% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1.R
\name{fig1}
\alias{fig1}
\title{Plot Fig 1}
\usage{
fig1(threads = c(1, 16), legend_position = "bottom", outfile = NULL)
}
\arguments{
\item{legend_position}{Where to position the legend in the figure}

\item{outfile}{Where to write the plot to file}
}
\description{
Plot Fig 1
}
