% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_feature}
\alias{get_feature}
\title{Get a feature from a regulatory gtf}
\usage{
get_feature(feature)
}
\arguments{
\item{feature}{The regulatory feature to extract ("promoter" or "enhancer")}
}
\description{
Get a feature from a regulatory gtf
}
