% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timediff.R
\name{get_timediff}
\alias{get_timediff}
\title{Get relative runtime differences}
\usage{
get_timediff(packages)
}
\arguments{
\item{packages}{The set of packages to read. Either \code{c("tabix")} or \code{c("BSseq", "biscuiteer", "query_all")}}

\item{outfile}{Where to write the plot to file}
}
\description{
Get relative runtime differences
}
