% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_exp_types}
\alias{join_exp_types}
\title{Join two benhmarks from sc and bulk}
\usage{
join_exp_types(sc, bulk)
}
\arguments{
\item{sc}{The sc benchmark data.table}

\item{bulk}{The sc benchmark data.table}
}
\description{
Join two benhmarks from sc and bulk
}
