% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_all.R
\name{plot_make_mat}
\alias{plot_make_mat}
\title{Plot iscream::make_bsseq_mat - Rsamtools::scanTabix benchmark}
\usage{
plot_make_mat(
  mat_benchmark,
  linewidth = 1,
  dot_size = 0.5,
  alpha = 0.8,
  outfile = NULL
)
}
\arguments{
\item{mat_benchmark}{The benchmark to plot (from \code{benchmark_make_bsseq_mat})}

\item{linewidth}{The linewidth to use for the mean runtime line}

\item{dot_size}{The size for the jitter runtime points}

\item{alpha}{The alpha level for the points}

\item{outfile}{Optional file to write the plot to}
}
\description{
Plot iscream::make_bsseq_mat - Rsamtools::scanTabix benchmark
}
