% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mem.R
\name{plot_mem}
\alias{plot_mem}
\title{Plot all package memory usage}
\usage{
plot_mem(peak = TRUE, sc = TRUE)
}
\arguments{
\item{peak}{Whether to plot peaks or memory usage over time}

\item{sc}{Whether to plot sc or bulk}
}
\description{
Plot all package memory usage
}
