% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_all.R
\name{plot_query_all}
\alias{plot_query_all}
\title{Plot iscream::query_all - Rsamtools::scanTabix benchmark}
\usage{
plot_query_all(
  query_all_benchmark,
  linewidth = 1,
  dot_size = 0.5,
  alpha = 0.8,
  outfile = NULL
)
}
\arguments{
\item{linewidth}{The linewidth to use for the mean runtime line}

\item{dot_size}{The size for the jitter runtime points}

\item{alpha}{The alpha level for the points}

\item{outfile}{Optional file to write the plot to}

\item{query_all}{The benchmark to plot (from \code{benchmark_query_all})}
}
\description{
Plot iscream::query_all - Rsamtools::scanTabix benchmark
}
