% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runtime_diff.R
\name{plot_runtime_diff}
\alias{plot_runtime_diff}
\title{Plot relative differences in runtime}
\usage{
plot_runtime_diff(sc = TRUE, file_count = 90)
}
\description{
Plot relative differences in runtime
}
