% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summarize_regions.R
\name{plot_summarize_regions}
\alias{plot_summarize_regions}
\title{Plot iscream::summarize_regions - Rsamtools::scanTabix benchmark}
\usage{
plot_summarize_regions(
  summarize_regions_benchmark,
  linewidth = 1,
  dot_size = 0.5,
  alpha = 0.8,
  outfile = NULL
)
}
\arguments{
\item{linewidth}{The linewidth to use for the mean runtime line}

\item{dot_size}{The size for the jitter runtime points}

\item{alpha}{The alpha level for the points}

\item{outfile}{Optional file to write the plot to}

\item{summarize_regions}{The benchmark to plot (from \code{benchmark_summarize_regions})}
}
\description{
Plot iscream::summarize_regions - Rsamtools::scanTabix benchmark
}
