% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tabix.R
\name{plot_tabix}
\alias{plot_tabix}
\title{Plot iscream::tabix - Rsamtools::scanTabix benchmark}
\usage{
plot_tabix(
  tabix_benchmark,
  linewidth = 1,
  dot_size = 1,
  alpha = 0.7,
  outfile = NULL
)
}
\arguments{
\item{tabix_benchmark}{The benchmark to plot (from \code{benchmark_tabix})}

\item{linewidth}{The linewidth to use for the mean runtime line}

\item{dot_size}{The size for the jitter runtime points}

\item{alpha}{The alpha level for the points}

\item{outfile}{Optional file to write the plot to}
}
\description{
Plot iscream::tabix - Rsamtools::scanTabix benchmark
}
