% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tabix.R
\name{plot_tabix_per_record}
\alias{plot_tabix_per_record}
\title{Plot iscream::tabix - Rsamtools::scanTabix benchmark per record}
\usage{
plot_tabix_per_record()
}
\arguments{
\item{sc_tabix}{The single-cell benchmark to plot (from \code{benchmark_tabix})}

\item{bulk_tabix}{The bulk benchmark to plot (from \code{benchmark_tabix})}

\item{linewidth}{The linewidth to use for the mean runtime line}

\item{dot_size}{The size for the jitter runtime points}

\item{alpha}{The alpha level for the points}

\item{outfile}{Optional file to write the plot to}
}
\description{
Plot iscream::tabix - Rsamtools::scanTabix benchmark per record
}
