% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tabix.R
\name{plot_tabix_shell}
\alias{plot_tabix_shell}
\title{Plot iscream::tabix - benchmarks against the shell tabix}
\usage{
plot_tabix_shell()
}
\description{
Plot iscream::tabix - benchmarks against the shell tabix
}
