% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_bulk}
\alias{read_bulk}
\title{Read bulk benchmarks}
\usage{
read_bulk()
}
\arguments{
\item{sc}{The sc benchmark data.table}

\item{bulk}{The sc benchmark data.table}
}
\description{
Read bulk benchmarks
}
