% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sc}
\alias{read_sc}
\title{Join two benhmarks from sc and bulk}
\usage{
read_sc()
}
\arguments{
\item{sc}{The sc benchmark data.table}

\item{bulk}{The sc benchmark data.table}
}
\description{
Join two benhmarks from sc and bulk
}
