% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_biscuiteer.R
\name{readbis}
\alias{readbis}
\title{Benchmark iscream::query_all on varying region, thread and file counts}
\usage{
readbis(bedfiles, gr, threads, merged)
}
\arguments{
\item{bedfiles}{A vector of bedfiles to run tests on}

\item{gr}{A GRanges of the regions}

\item{threads}{The number of threads to run mclapply with}

\item{merged}{Whether the files are CG merged}
}
\description{
Benchmark iscream::query_all on varying region, thread and file counts
}
