% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beds.R
\name{search_beds}
\alias{search_beds}
\title{Get a vector of bedfiles based on experiment type, aligner and dataset name.
If only one match is found, the vector is returned, else matching datasets
are displayed}
\usage{
search_beds(
  experiment_type = NULL,
  aligner_name = NULL,
  dataset_name = NULL,
  mergecg = TRUE
)
}
\arguments{
\item{experiment_type}{The experiment type (eg: "sc", or "bulk")}

\item{aligner_name}{The aligner name ("biscuit", "bismark", "bsbolt")}

\item{dataset_name}{The name of the dataset}

\item{mergecg}{Whether to return BISCUIT mergecg BEDs or not}
}
\description{
Get a vector of bedfiles based on experiment type, aligner and dataset name.
If only one match is found, the vector is returned, else matching datasets
are displayed
}
