% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_datasets.R
\name{setup_datasets}
\alias{setup_datasets}
\title{Get a table of the datasets provided in the config}
\usage{
setup_datasets(iscream_run_conf)
}
\arguments{
\item{iscream_run_conf}{the RcppTOML struct containing the config}
}
\description{
Get a table of the datasets provided in the config
}
