% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_summarize_regions.R
\name{benchmark_summarize_regions}
\alias{benchmark_summarize_regions}
\title{Benchmark iscream::summarize regions on varying region, thread and file counts}
\usage{
benchmark_summarize_regions(
  bedfiles,
  regions,
  min_iterations = 3,
  max_iterations = 10,
  raw = TRUE,
  n_regions = c(1, 100, 500, 1000, 5000, 10000),
  n_files = round(10^c(0, 1, 1.4, 1.7, 1.875, 2)),
  n_threads = 2^c(0, 1, 2, 3, 4),
  outfile = NULL
)
}
\arguments{
\item{bedfiles}{A vector of bedfiles to run tests on}

\item{regions}{A vector of regions to query from}

\item{min_iterations}{The min iterations for bench::mark to run}

\item{max_iterations}{The max iterations for bench::mark to run}

\item{n_regions}{A vector of region counts to benchmark}

\item{n_files}{A vector of file counts to benchmark}

\item{n_threads}{A vector of thread counts to benchmark}

\item{outfile}{Optional file to write the benchmark to}

\item{threads}{The number of threads to run on}
}
\description{
Benchmark iscream::summarize regions on varying region, thread and file counts
}
