% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_summarize_regions.R
\name{benchmark_summarize_regions}
\alias{benchmark_summarize_regions}
\title{Benchmark iscream::summarize regions on varying region, thread and file counts}
\usage{
benchmark_summarize_regions(
  bedfile,
  regions_file,
  sfun,
  min_iterations = 3,
  max_iterations = 10,
  n_regions = 30000,
  outfile = NULL
)
}
\arguments{
\item{bedfile}{A bedfile to run test on}

\item{regions_file}{A bed file containing the regions of interest}

\item{sfun}{The functions to use}

\item{min_iterations}{The min iterations for bench::mark to run}

\item{max_iterations}{The max iterations for bench::mark to run}

\item{n_regions}{A vector of region counts to benchmark}

\item{outfile}{Optional file to write the benchmark to}

\item{n_files}{A vector of file counts to benchmark}

\item{n_threads}{A vector of thread counts to benchmark}
}
\description{
Benchmark iscream::summarize regions on varying region, thread and file counts
}
